@@ -23,8 +23,9 @@
         <!--<editor-fold desc="params">-->
         <coverageSurefireArgLine></coverageSurefireArgLine>
         <docker-target-image>lermitage-manon:${project.version}</docker-target-image>
-        <maven.compiler.source>11</maven.compiler.source>
-        <maven.compiler.target>11</maven.compiler.target>
+        <java.version>17</java.version>
+        <maven.compiler.source>${java.version}</maven.compiler.source>
+        <maven.compiler.target>${java.version}</maven.compiler.target>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
         <start-class>manon.Application</start-class>
@@ -38,34 +39,35 @@
         <docker-base-image>eclipse-temurin:17.0.5_8-jre</docker-base-image>
 
         <!-- Override spring-boot-dependencies versions: -->
-        <!-- - version provided by spring-boot-dependencies is a bit old (v3). -->
-        <rest-assured.version>5.2.0</rest-assured.version> <!-- https://github.com/rest-assured/rest-assured/blob/master/changelog.txt -->
+        <!-- - version provided by spring-boot-dependencies may be a bit old. -->
+        <rest-assured.version>5.3.0</rest-assured.version> <!-- https://github.com/rest-assured/rest-assured/blob/master/changelog.txt -->
         <!-- - to let versions-maven-plugin handle version. Don't forget to update spring-boot-starter-parent manually! -->
-        <spring-boot.version>2.7.5</spring-boot.version>
+        <spring-boot.version>3.0.0</spring-boot.version>
 
-        <archunit.version>1.0.0</archunit.version>                  <!-- https://github.com/TNG/ArchUnit/releases -->
-        <equalsverifier.version>3.10.1</equalsverifier.version>      <!-- https://github.com/jqno/equalsverifier/blob/main/CHANGELOG.md -->
-        <greenmail.version>1.6.11</greenmail.version>                <!-- https://github.com/greenmail-mail-test/greenmail/releases -->
-        <guava.version>31.1-jre</guava.version>                      <!-- https://github.com/google/guava/releases -->
-        <javafaker.version>1.0.2</javafaker.version>                 <!-- https://github.com/DiUS/java-faker/tags -->
+        <archunit.version>1.0.1</archunit.version> <!-- https://github.com/TNG/ArchUnit/releases -->
+        <datafaker.version>1.7.0</datafaker.version> <!-- https://github.com/datafaker-net/datafaker/tags -->
+        <equalsverifier.version>3.11.1</equalsverifier.version> <!-- https://github.com/jqno/equalsverifier/blob/main/CHANGELOG.md -->
+        <greenmail.version>2.0.0-alpha-2</greenmail.version> <!-- https://github.com/greenmail-mail-test/greenmail/releases -->
+        <guava.version>31.1-jre</guava.version> <!-- https://github.com/google/guava/releases -->
         <javax.activation-activation.version>1.1.1</javax.activation-activation.version>
         <jetbrains-annotations.version>23.0.0</jetbrains-annotations.version> <!-- https://github.com/JetBrains/java-annotations/blob/master/CHANGELOG.md -->
-        <jjwt.version>0.11.5</jjwt.version>                          <!-- https://github.com/jwtk/jjwt/blob/master/CHANGELOG.md -->
-        <json-unit.version>2.36.0</json-unit.version>                <!-- https://github.com/lukas-krecan/JsonUnit#release-notes -->
+        <jjwt.version>0.11.5</jjwt.version> <!-- https://github.com/jwtk/jjwt/blob/master/CHANGELOG.md -->
+        <json-unit.version>2.36.0</json-unit.version> <!-- https://github.com/lukas-krecan/JsonUnit#release-notes -->
         <lombok-mapstruct-binding.version>0.2.0</lombok-mapstruct-binding.version>
-        <mapstruct.version>1.5.3.Final</mapstruct.version>           <!-- https://github.com/mapstruct/mapstruct/releases -->
-        <springdoc.version>1.6.12</springdoc.version>                <!-- https://springdoc.org/#migrating-from-springfox -->
+        <mapstruct.version>1.5.3.Final</mapstruct.version> <!-- https://github.com/mapstruct/mapstruct/releases -->
+        <springdoc.version>2.0.0</springdoc.version> <!-- https://springdoc.org/v2/ https://springdoc.org/#migrating-from-springfox -->
 
         <apt-maven-plugin.version>1.1.3</apt-maven-plugin.version>
         <findbugs-jsr305.version>3.0.2</findbugs-jsr305.version>
         <findsecbugs-plugin.version>1.12.0</findsecbugs-plugin.version>
+        <git-commit-id-maven-plugin.version>5.0.0</git-commit-id-maven-plugin.version>
         <jacoco-maven-plugin.version>0.8.8</jacoco-maven-plugin.version> <!-- https://www.jacoco.org/jacoco/trunk/doc/changes.html (lib used by plugin) -->
-        <jib-maven-plugin.version>3.3.0</jib-maven-plugin.version> <!-- https://github.com/GoogleContainerTools/jib/blob/master/jib-maven-plugin/CHANGELOG.md -->
+        <jib-maven-plugin.version>3.3.1</jib-maven-plugin.version> <!-- https://github.com/GoogleContainerTools/jib/blob/master/jib-maven-plugin/CHANGELOG.md -->
         <oga-maven-plugin.version>1.8.0</oga-maven-plugin.version>
-        <owasp-dependency-check-plugin.version>7.3.0</owasp-dependency-check-plugin.version>
+        <owasp-dependency-check-plugin.version>7.3.2</owasp-dependency-check-plugin.version>
         <pitest-junit5-plugin.version>1.1.0</pitest-junit5-plugin.version>
-        <pitest-maven.version>1.9.8</pitest-maven.version>
-        <spotbugs-maven-plugin.version>4.7.2.1</spotbugs-maven-plugin.version>
+        <pitest-maven.version>1.9.11</pitest-maven.version>
+        <spotbugs-maven-plugin.version>4.7.3.0</spotbugs-maven-plugin.version>
         <versions-maven-plugin.version>2.12.0</versions-maven-plugin.version> <!-- https://github.com/mojohaus/versions-maven-plugin/releases -->
     </properties>
 
