@@ -2,16 +2,15 @@ package manon.batch.listener;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.batch.core.JobExecution;
-import org.springframework.batch.core.listener.JobExecutionListenerSupport;
+import org.springframework.batch.core.JobExecutionListener;
 import org.springframework.stereotype.Component;
 
 @Component
 @Slf4j
-public class JobListener extends JobExecutionListenerSupport {
+public class JobListener implements JobExecutionListener {
 
     @Override
     public void afterJob(JobExecution jobExecution) {
-        super.afterJob(jobExecution);
         log.info("job {} ended in state {}",
             jobExecution.getJobInstance().getJobName(),
             jobExecution.getStatus());
