@@ -1,12 +1,11 @@
 package manon.model.user.form;
 
+import jakarta.validation.constraints.NotNull;
+import jakarta.validation.constraints.Size;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-
-import javax.validation.constraints.NotNull;
-import javax.validation.constraints.Size;
 
 import static lombok.AccessLevel.PRIVATE;
 import static manon.document.user.UserEntity.Validation.PASSWORD_MAX_LENGTH;
