@@ -1,7 +1,6 @@
 package manon.repository.user;
 
 import manon.document.user.FriendshipEntity;
-import manon.repository.WorkaroundUntilHibernate6;
 import manon.util.ExistForTesting;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
@@ -13,9 +12,7 @@ import java.util.List;
 import java.util.stream.Stream;
 
 @Repository
-public interface FriendshipRepository
-    extends JpaRepository<FriendshipEntity, Long>,
-    WorkaroundUntilHibernate6<FriendshipEntity> {
+public interface FriendshipRepository extends JpaRepository<FriendshipEntity, Long> {
 
     /** Find a friendship between two users. */
     @Query("select count(f) from Friendship f " +
