@@ -4,7 +4,6 @@ import manon.document.user.UserEntity;
 import manon.document.user.UserIdProjection;
 import manon.document.user.UserVersionProjection;
 import manon.model.user.RegistrationState;
-import manon.repository.WorkaroundUntilHibernate6;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
@@ -15,10 +14,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface UserRepository
-    extends JpaRepository<UserEntity, Long>,
-    QuerydslPredicateExecutor<UserEntity>,
-    WorkaroundUntilHibernate6<UserEntity> {
+public interface UserRepository extends JpaRepository<UserEntity, Long>, QuerydslPredicateExecutor<UserEntity> {
 
     Optional<UserEntity> findByUsername(String username);
 
